Add the file upload component
This was an interesting thing to develop!  Ember doesn't appear to
support observing file inputs natively, so I had to attach a handler
after the component was inserted.  At that point, it was a simple matter
of using javascript's file API to grab the image data as a url.  Then, I
send the components main action wth the image data to the parent
component/controller.