rename the file-upload component to image-upload
This is really built for an image uploader, so I renamed it as such.

It might be better to also pass an element selector to this component as
the 'preview' element, and contain all of the image upload and display
logic in the component.  That's next.

Also removed a wayward debugger.