Reorganize the actions area
I want the lists to fill from the top.  I was using inline-block, which
anchors things to the baseline though, and inserts spaces between
inline-block elements with whitespace between them in the document.
This is all wrong, so instead, I changed `display: inline-block` to
`float: left`, and now things fit nicely.

I also moved the reset button to the top, and centered it so it wasn't
moving around. I hate it when UI elements move around!